Fix the issue caused by organizing aliases.
When I'm writing functions, I occasionally need to handle unused variables, but I find that organizing aliases still returns results. This causes a lot of interference, especially when there's a bug in nvim, as pressing keys quickly can mess up my code.

We should not return the results of organizing aliases when the cursor is not on a line with an alias or when the current token is not an alias.